validaciones y readme con explicacion de ejecucion

+++ app/resources/task.py
@@ -2,6 +2,7 @@ from flask import redirect, render_template, request, url_for, session, flash
 
 from app.models.task import Task
 from app.models.folder import Folder
+from app.models.user import User
 from app.helpers.forms import FormTask,FormTaskState
 from app.helpers.auth import authenticated
 from app.helpers.validates import sanitizar_input
@@ -13,13 +14,22 @@ from app import db
 def index():
     """
     """
+    if not authenticated(session):
+       return redirect(url_for("home"))
+    
     return render_template("task/index.html")
 
 def new(folder_id):
     """
-    """
+    """ 
     if not authenticated(session):
-      return redirect(url_for("home"))
+       return redirect(url_for("home"))
+    
+    folders = User.with_id(session["id"]).folders
+    folder = Folder.with_id(folder_id)
+    if not folder in folders:
+       return redirect(request.referrer)
+
     form = FormTask()
     form.folder_id.data = folder_id
     
@@ -28,8 +38,18 @@ def new(folder_id):
 def create():
     """
     """
-    form = FormTask()
+    if not authenticated(session):
+       return redirect(url_for("home"))
+    
+    folders = User.with_id(session["id"]).folders
     folder = Folder.with_id(request.form["folder_id"])
+    if not folder in folders:
+       
+       return redirect(url_for("home"))
+
+
+    form = FormTask()
+   
     if any(task.name == form.name.data for task in folder.tasks):
          flash("The name is alredy in use")
          return redirect(request.referrer)
@@ -44,18 +64,39 @@ def create():
 def show(task_id):
     """
     """
-
+    if not authenticated(session):
+       return redirect(url_for("home"))
+    
+    folders = User.with_id(session["id"]).folders
     task = Task.with_id(task_id)
+    if not task:
+       return redirect(url_for("home"))
+
+    folder = Folder.with_id(task.folder_id)
+    if not folder in folders:
+       return redirect(url_for("home"))
+
+
     form = FormTaskState()
     return render_template("task/show.html", task= task,form=form)
 
 def edit(task_id):
-    if not authenticated(session):
-      return redirect(url_for("home"))
-
     "validar que sea del usuario logeado"
     """
     """
+    if not authenticated(session):
+       return redirect(url_for("home"))
+    
+    folders = User.with_id(session["id"]).folders
+    task=Task.with_id(task_id)
+    if not task:
+       return redirect(url_for("home"))
+
+    folder = Folder.with_id(task.folder_id)
+    if not folder in folders:
+       return redirect(url_for("home"))
+
+
     task = Task.with_id(task_id)
     form = FormTask()
     form.name.data = task.name
@@ -65,10 +106,18 @@ def edit(task_id):
 def update():
     """
     """
+    if not authenticated(session):
+       return redirect(url_for("home"))
+    
+    folders = User.with_id(session["id"]).folders
+    task=Task.with_id(request.form["task_id"])
+    if not task:
+       return redirect(url_for("home"))
 
+    folder = Folder.with_id(task.folder_id)
+    if not folder in folders:
+       return redirect(url_for("home"))
 
-
-    task = Task.with_id(request.form["folder_id"])
     form = FormTaskState()
     task.update(request.form)
 
@@ -77,8 +126,18 @@ def update():
 def delete():
     """
     """
-    task = Task.with_id(request.form["task_id"])
-    folder = task.folder
+    if not authenticated(session):
+       return redirect(url_for("home"))
+    
+    folders = User.with_id(session["id"]).folders
+    task=Task.with_id(request.form["task_id"])
+    if not task:
+       return redirect(url_for("home"))
+
+    folder = Folder.with_id(task.folder_id)
+    if not folder in folders:
+       return redirect(request.referrer)
+  
 
     task.delete() 
 
@@ -87,6 +146,18 @@ def delete():
 def completed():
     """
     """ 
+    if not authenticated(session):
+       return redirect(url_for("home"))
+    
+    folders = User.with_id(session["id"]).folders
+    task=Task.with_id(request.form["task_id"])
+    if not task:
+       return redirect(url_for("home"))
+
+    folder = Folder.with_id(task.folder_id)
+    if not folder in folders:
+       return redirect(request.referrer)
+  
     task=Task.with_id(request.form["task_id"])
     task.completed()
     return redirect(request.referrer)
@@ -95,7 +166,19 @@ def completed():
 def inprogress():
     """
     """
+    if not authenticated(session):
+       return redirect(url_for("home"))
+    
+    folders = User.with_id(session["id"]).folders
     task=Task.with_id(request.form["task_id"])
+    if not task:
+       return redirect(url_for("home"))
+
+    folder = Folder.with_id(task.folder_id)
+    if not folder in folders:
+       return redirect(request.referrer)
+
+
     task.inprogress()
     return redirect(request.referrer)
